+++ gwcs/api.py
@@ -135,9 +135,10 @@ class GWCSAPIMixin(BaseHighLevelWCS, BaseLowLevelWCS):
         world_arrays = self._add_units_input(world_arrays, self.backward_transform, self.output_frame)
 
         result = self.invert(*world_arrays, with_units=False)
-        result = self._out_of_bounding_box_to_nan(result, world_arrays)
+        result = self._remove_quantity_output(result, self.input_frame)
 
-        return self._remove_quantity_output(result, self.input_frame)
+        return self._out_of_bounding_box_to_nan(result, world_arrays)
+
 
     def world_to_array_index_values(self, *world_arrays):
         """